Relocate json-related method to JsonUtilities file
