+++ src/lib/support/tests/TestJsonToTlvToJson.cpp
@@ -20,6 +20,7 @@
 #include <app/data-model/Encode.h>
 #include <lib/support/UnitTestRegistration.h>
 #include <lib/support/jsontlv/JsonToTlv.h>
+#include <lib/support/jsontlv/JsonUtilities.h>
 #include <lib/support/jsontlv/TlvToJson.h>
 #include <nlunit-test.h>
