+++ src/lib/support/tests/TestTlvToJson.cpp
@@ -19,6 +19,7 @@
 #include <app/data-model/Decode.h>
 #include <app/data-model/Encode.h>
 #include <lib/support/UnitTestRegistration.h>
+#include <lib/support/jsontlv/JsonUtilities.h>
 #include <lib/support/jsontlv/TlvToJson.h>
 #include <nlunit-test.h>
 #include <system/SystemPacketBuffer.h>
